requirements.txt
pytz
astral
Pillow
freetype-py
